+++ tests/benchdnn/conv/ref_conv.cpp
@@ -50,17 +50,18 @@ void compute_ref_bwd_w(const prb_t *p, dnn_mem_t &src_m, dnn_mem_t &diff_wei_m,
 
 void compute_ref_direct_fwd(const prb_t *p, dnn_mem_t &src_m,
         dnn_mem_t &wei_m, dnn_mem_t &bia_m, dnn_mem_t &dst_m) {
-    auto ker = [&](float &d, int g, int mb, int oc, int od, int oh, int ow) {
-        for (int ic = 0; ic < p->ic/p->g; ++ic) {
-            for (int kd = 0; kd < p->kd; ++kd) {
-                const int id = od * p->sd - p->pd + kd * (p->dd + 1);
+    auto ker = [&](float &d, int64_t g, int64_t mb,
+            int64_t oc, int64_t od, int64_t oh, int64_t ow) {
+        for (int64_t ic = 0; ic < p->ic/p->g; ++ic) {
+            for (int64_t kd = 0; kd < p->kd; ++kd) {
+                const int64_t id = od * p->sd - p->pd + kd * (p->dd + 1);
                 if (id < 0 || id >= p->id) continue;
-                for (int kh = 0; kh < p->kh; ++kh) {
-                    const int ih = oh * p->sh - p->ph + kh * (p->dh + 1);
+                for (int64_t kh = 0; kh < p->kh; ++kh) {
+                    const int64_t ih = oh * p->sh - p->ph + kh * (p->dh + 1);
                     if (ih < 0 || ih >= p->ih) continue;
 
-                    for (int kw = 0; kw < p->kw; ++kw) {
-                        const int iw = ow * p->sw - p->pw + kw * (p->dw + 1);
+                    for (int64_t kw = 0; kw < p->kw; ++kw) {
+                        const int64_t iw = ow * p->sw - p->pw + kw * (p->dw + 1);
                         if (iw < 0 || iw >= p->iw) continue;
 
                         size_t src_off = src_off_f(p, mb, g, ic, id, ih, iw);
@@ -73,7 +74,7 @@ void compute_ref_direct_fwd(const prb_t *p, dnn_mem_t &src_m,
         }
     };
 
-    auto maybe_scale = [&](float &d, int oc) {
+    auto maybe_scale = [&](float &d, int64_t oc) {
         if (!p->attr.oscale.is_def()) {
             using policy_t = attr_t::scale_t::policy_t;
             const auto &s = p->attr.oscale;
@@ -116,7 +117,7 @@ void compute_ref_direct_fwd(const prb_t *p, dnn_mem_t &src_m,
     };
 
     mkldnn::impl::parallel_nd(p->g, p->mb, p->oc / p->g, p->od, p->oh, p->ow,
-        [&](int g, int mb, int oc, int od, int oh, int ow) {
+        [&](int64_t g, int64_t mb, int64_t oc, int64_t od, int64_t oh, int64_t ow) {
             const size_t dst_off = dst_off_f(p, mb, g, oc, od, oh, ow);
             float &dst = ((float*)dst_m)[dst_off];
 
@@ -142,12 +143,12 @@ void compute_ref_direct_bwd_d(const prb_t *p, dnn_mem_t &diff_src_m,
     const bool fast = MAX2(p->kh, p->kw) <= precompute_size;
 
     /* pre-computes arrays of oh(ow) and kh(kw) for traversing in kernel */
-    auto precompute_ok = [](int i, int O, int K, int S, int P, int D,
-            int &num, int *_o, int *_k) {
+    auto precompute_ok = [](int64_t i, int64_t O, int64_t K, int64_t S,
+            int64_t P, int64_t D, int64_t &num, int64_t *_o, int64_t *_k) {
         assert(K <= precompute_size);
         num = 0;
-        for (int k = 0; k < K; ++k) {
-            int o = i - k * (D + 1) + P;
+        for (int64_t k = 0; k < K; ++k) {
+            int64_t o = i - k * (D + 1) + P;
             if (o < 0 || o % S) continue;
             o /= S;
             if (o >= O) continue;
@@ -157,18 +158,19 @@ void compute_ref_direct_bwd_d(const prb_t *p, dnn_mem_t &diff_src_m,
         }
     };
 
-    auto ker_fast = [&](float &ds, int g, int mb, int ic, int id, int ih, int iw) {
-        int kd[precompute_size], od[precompute_size], num_d;
-        int kh[precompute_size], oh[precompute_size], num_h;
-        int kw[precompute_size], ow[precompute_size], num_w;
+    auto ker_fast = [&](float &ds, int64_t g, int64_t mb,
+            int64_t ic, int64_t id, int64_t ih, int64_t iw) {
+        int64_t kd[precompute_size], od[precompute_size], num_d;
+        int64_t kh[precompute_size], oh[precompute_size], num_h;
+        int64_t kw[precompute_size], ow[precompute_size], num_w;
         precompute_ok(id, p->od, p->kd, p->sd, p->pd, p->dd, num_d, od, kd);
         precompute_ok(ih, p->oh, p->kh, p->sh, p->ph, p->dh, num_h, oh, kh);
         precompute_ok(iw, p->ow, p->kw, p->sw, p->pw, p->dw, num_w, ow, kw);
 
-        for (int oc = 0; oc < p->oc/p->g; ++oc) {
-            for (int d = 0; d < num_d; ++d) {
-                for (int h = 0; h < num_h; ++h) {
-                    for (int w = 0; w < num_w; ++w) {
+        for (int64_t oc = 0; oc < p->oc/p->g; ++oc) {
+            for (int64_t d = 0; d < num_d; ++d) {
+                for (int64_t h = 0; h < num_h; ++h) {
+                    for (int64_t w = 0; w < num_w; ++w) {
 
                         size_t dst_off = dst_off_f(p, mb, g, oc, od[d], oh[h], ow[w]);
                         size_t wei_off = wei_off_f(p, g, oc, ic, kd[d], kh[h], kw[w]);
@@ -180,21 +182,22 @@ void compute_ref_direct_bwd_d(const prb_t *p, dnn_mem_t &diff_src_m,
         }
     };
 
-    auto ker = [&](float &ds, int g, int mb, int ic, int id, int ih, int iw) {
-        for (int oc = 0; oc < p->oc/p->g; ++oc) {
-            for (int kd = 0; kd < p->kd; ++kd) {
-                int od = id - kd * (p->dd + 1) + p->pd;
+    auto ker = [&](float &ds, int64_t g, int64_t mb,
+            int64_t ic, int64_t id, int64_t ih, int64_t iw) {
+        for (int64_t oc = 0; oc < p->oc/p->g; ++oc) {
+            for (int64_t kd = 0; kd < p->kd; ++kd) {
+                int64_t od = id - kd * (p->dd + 1) + p->pd;
                 if (od < 0 || od % p->sd) continue;
                 od /= p->sd;
                 if (od >= p->od) continue;
-                for (int kh = 0; kh < p->kh; ++kh) {
-                    int oh = ih - kh * (p->dh + 1) + p->ph;
+                for (int64_t kh = 0; kh < p->kh; ++kh) {
+                    int64_t oh = ih - kh * (p->dh + 1) + p->ph;
                     if (oh < 0 || oh % p->sh) continue;
                     oh /= p->sh;
                     if (oh >= p->oh) continue;
 
-                    for (int kw = 0; kw < p->kw; ++kw) {
-                        int ow = iw - kw * (p->dw + 1) + p->pw;
+                    for (int64_t kw = 0; kw < p->kw; ++kw) {
+                        int64_t ow = iw - kw * (p->dw + 1) + p->pw;
                         if (ow < 0 || ow % p->sw) continue;
                         ow /= p->sw;
                         if (ow >= p->ow) continue;
@@ -209,7 +212,7 @@ void compute_ref_direct_bwd_d(const prb_t *p, dnn_mem_t &diff_src_m,
         }
     };
 
-    auto maybe_scale = [&](float &ds, int ic) {
+    auto maybe_scale = [&](float &ds, int64_t ic) {
         if (!p->attr.oscale.is_def()) {
             using policy_t = attr_t::scale_t::policy_t;
             const auto &s = p->attr.oscale;
@@ -253,7 +256,7 @@ void compute_ref_direct_bwd_d(const prb_t *p, dnn_mem_t &diff_src_m,
     };
 
     mkldnn::impl::parallel_nd(p->g, p->mb, p->ic / p->g, p->id, p->ih, p->iw,
-        [&](int g, int mb, int ic, int id, int ih, int iw) {
+        [&](int64_t g, int64_t mb, int64_t ic, int64_t id, int64_t ih, int64_t iw) {
             size_t src_off = src_off_f(p, mb, g, ic, id, ih, iw);
             float &ds = ((float*)diff_src_m)[src_off];
             float conv_res = 0;
@@ -276,26 +279,27 @@ void compute_ref_direct_bwd_d(const prb_t *p, dnn_mem_t &diff_src_m,
 
 void compute_ref_bwd_weights(const prb_t *p, dnn_mem_t &src_m,
         dnn_mem_t &diff_wei_m, dnn_mem_t &diff_dst_m) {
-    auto compute_bounds = [](int I, int O, int k, int S, int P, int D,
-            int &o_s, int &o_e) {
+    auto compute_bounds = [](int64_t I, int64_t O, int64_t k, int64_t S,
+            int64_t P, int64_t D, int64_t &o_s, int64_t &o_e) {
         const float tmp = P - k * (D + 1);
         o_s = MAX2(0, ceilf(tmp / S));
         o_e = MIN2(O, ceilf((I + tmp) / S));
     };
 
-    auto ker = [&](float &dw, int g, int oc, int ic, int kd, int kh, int kw) {
-        int od_s, od_e, oh_s, oh_e, ow_s, ow_e;
+    auto ker = [&](float &dw, int64_t g, int64_t oc, int64_t ic,
+            int64_t kd, int64_t kh, int64_t kw) {
+        int64_t od_s, od_e, oh_s, oh_e, ow_s, ow_e;
         compute_bounds(p->id, p->od, kd, p->sd, p->pd, p->dd, od_s, od_e);
         compute_bounds(p->ih, p->oh, kh, p->sh, p->ph, p->dh, oh_s, oh_e);
         compute_bounds(p->iw, p->ow, kw, p->sw, p->pw, p->dw, ow_s, ow_e);
 
-        for (int mb = 0; mb < p->mb; ++mb) {
-            for (int od = od_s; od < od_e; ++od) {
-            for (int oh = oh_s; oh < oh_e; ++oh) {
-            for (int ow = ow_s; ow < ow_e; ++ow) {
-                const int id = od * p->sd - p->pd + kd * (p->dd + 1);
-                const int ih = oh * p->sh - p->ph + kh * (p->dh + 1);
-                const int iw = ow * p->sw - p->pw + kw * (p->dw + 1);
+        for (int64_t mb = 0; mb < p->mb; ++mb) {
+            for (int64_t od = od_s; od < od_e; ++od) {
+            for (int64_t oh = oh_s; oh < oh_e; ++oh) {
+            for (int64_t ow = ow_s; ow < ow_e; ++ow) {
+                const int64_t id = od * p->sd - p->pd + kd * (p->dd + 1);
+                const int64_t ih = oh * p->sh - p->ph + kh * (p->dh + 1);
+                const int64_t iw = ow * p->sw - p->pw + kw * (p->dw + 1);
 
                 size_t src_off = src_off_f(p, mb, g, ic, id, ih, iw);
                 size_t dst_off = dst_off_f(p, mb, g, oc, od, oh, ow);
@@ -309,7 +313,7 @@ void compute_ref_bwd_weights(const prb_t *p, dnn_mem_t &src_m,
 
     mkldnn::impl::parallel_nd(
         p->g, p->oc / p->g, p->ic / p->g, p->kd, p->kh, p->kw,
-        [&](int g, int oc, int ic, int kd, int kh, int kw) {
+        [&](int64_t g, int64_t oc, int64_t ic, int64_t kd, int64_t kh, int64_t kw) {
                 size_t wei_off = wei_off_f(p, g, oc, ic, kd, kh, kw);
                 float &dw = ((float*)diff_wei_m)[wei_off];
                 dw = 0;
@@ -320,14 +324,14 @@ void compute_ref_bwd_weights(const prb_t *p, dnn_mem_t &src_m,
 
 void compute_ref_bwd_bias(const prb_t *p, dnn_mem_t &diff_bia_m,
     dnn_mem_t &diff_dst_m) {
-    mkldnn::impl::parallel_nd(p->g, p->oc / p->g, [&](int g, int oc) {
+    mkldnn::impl::parallel_nd(p->g, p->oc / p->g, [&](int64_t g, int64_t oc) {
        size_t bia_off = bia_off_f(p, g, oc);
        double sum = 0;
 
-       for (int mb = 0; mb < p->mb; ++mb)
-       for (int od = 0; od < p->od; ++od)
-       for (int oh = 0; oh < p->oh; ++oh)
-       for (int ow = 0; ow < p->ow; ++ow)
+       for (int64_t mb = 0; mb < p->mb; ++mb)
+       for (int64_t od = 0; od < p->od; ++od)
+       for (int64_t oh = 0; oh < p->oh; ++oh)
+       for (int64_t ow = 0; ow < p->ow; ++ow)
        {
            size_t dst_off = dst_off_f(p, mb, g, oc, od, oh, ow);
            sum += ((float*)diff_dst_m)[dst_off];
